Bring back missing custom function
